requirements.txt
asgiref==3.3.4
backoff==1.10.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
Deprecated==1.2.12
Django==3.2.9
django-bulma==0.8.2
django-loki==0.1.4
django-prometheus==2.1.0
djangorestframework==3.12.4
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==3.2
mysql-connector-python==8.0.27
mysqlclient==2.0.3
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-exporter-jaeger-proto-grpc==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-exporter-otlp==1.7.1
opentelemetry-exporter-otlp-proto-grpc==1.7.1
opentelemetry-exporter-otlp-proto-http==1.7.1
opentelemetry-exporter-zipkin==1.7.1
opentelemetry-exporter-zipkin-json==1.7.1
opentelemetry-exporter-zipkin-proto-http==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-asgi==0.26b1
opentelemetry-instrumentation-dbapi==0.26b1
opentelemetry-instrumentation-django==0.26b1
opentelemetry-instrumentation-logging==0.26b1
opentelemetry-instrumentation-mysql==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-sqlite3==0.26b1
opentelemetry-instrumentation-urllib==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-propagator-b3==1.7.1
opentelemetry-proto==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
prometheus-client==0.11.0
protobuf==3.17.3
python-json-logger==2.0.2
python-logging-loki==0.3.1
pytz==2021.1
requests==2.26.0
rfc3339==6.2
six==1.16.0
sqlparse==0.4.1
thrift==0.13.0
urllib3==1.26.6
uWSGI==2.0.19.1
wrapt==1.12.1
